requirements.txt
streamlit==1.24.0
pytube==12.1.0
youtube-transcript-api==0.6.2
python-pptx==0.6.21
gtts==2.3.0
moviepy==1.0.3
transformers==4.31.0
torch==2.0.1
